PetAlreadyExistsExceptionの修正と計算ロジックをServiceへ移動
