+++ src/main/java/com/u_1/copypet/Service/PetService.java
@@ -1,8 +1,11 @@
 package com.u_1.copypet.Service;
 
 import com.u_1.copypet.Controller.Response.PetAlreadyExistsException;
+import com.u_1.copypet.Entity.Enum.ActivityLevel;
+import com.u_1.copypet.Entity.Enum.Gender;
 import com.u_1.copypet.Entity.Pet;
 import com.u_1.copypet.Mapper.PetMapper;
+import java.util.Optional;
 import org.springframework.stereotype.Service;
 
 
@@ -20,12 +23,39 @@ public class PetService {
     return pet;
   }
 
-  public Pet findPetByUserId(int userId) {
-    petMapper.findPetByUserId(userId).ifPresent(pet -> {
-      throw new PetAlreadyExistsException("Pet already exists with user id: " + userId);
-    });
-    return null;//petが存在しないければ進行
+  //修正箇所
+  public Optional<Pet> findPetByUserId(int userId) {
+    petMapper.findPetByUserId(userId)
+        .ifPresent(findPetByUserId -> {
+          throw new PetAlreadyExistsException("Pet already exists with user id: " + userId);
+        });
+    return Optional.empty();
   }
+
+
+  // bmrを計算するメソッド
+  public double calculateBmr(Gender gender, int age, double height, double weight) {
+    if (gender == Gender.MALE) {
+      return Math.round((88.362 + (13.397 * weight) + (4.799 * height) - (5.677 * age)) * 100)
+          / 100.0;
+    } else { // Gender.FEMALE
+      return Math.round((447.593 + (9.247 * weight) + (3.098 * height) - (4.330 * age)) * 100)
+          / 100.0;
+    }
+  }
+
+  // 1日のエネルギー必要量を計算するメソッド
+  public double calculateDailyEnergyRequirement(double bmr, ActivityLevel activityLevel) {
+
+    return switch (activityLevel) {
+      case LOW -> bmr * 1.2;
+      case MEDIUM -> bmr * 1.75;
+      case HIGH -> bmr * 2.0;
+      default -> 0;
+    };
+  }
+
 }
+
